Split some actors to separated headers
