+++ .gitignore
@@ -11,3 +11,4 @@ Build
 Plugins/StoryGraphPlugin/Binaries
 Saved
 .vs
+.idea
