+++ Plugins/StoryGraphPlugin/Source/StoryGraphPluginRuntime/Public/ObjectRecord.h
@@ -1,8 +1,5 @@
 #pragma once
 #include "Object.h"
-#include "Interface.h"
-#include "ObjectRecord.generated.h"
-
 
 struct FSaveArchiveHeader
 {
@@ -27,25 +24,6 @@ struct FObjectRecord
 	TArray<FObjectRecord> InternallyObjects;
 	void Load(UObject* Object);
 };
-
-
-UINTERFACE()
-class USaveObject_StoryGraph : public UInterface
-{
-	GENERATED_BODY()
-};
-
-class ISaveObject_StoryGraph
-{
-	GENERATED_BODY()
-public:
-	bool IsLoaded;
-public:
-	virtual void GetInternallySaveObjects(TArray<UObject*>& Objects, int WantedObjectsNum)
-	{
-	} //if we load object we take function number objects witch we need
-};
-
 
 FORCEINLINE FArchive& operator<<(FArchive& Ar, FSaveArchiveHeader& ArchiveHeader)
 {
